refactor(devices): share findDrumPad with pad-level callers
Pad-level operations need the DrumPad itself, not its chain — Live's
delete_all_chains and copy_pad are silent no-ops when aimed at a chain.
delete had the only lookup; move it beside the other pad navigation so
the next caller doesn't grow a second copy.
